pyproject.toml
[project]
name = "openai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "isort>=6.0.1",
    "scipy>=1.15.3",
    "sounddevice>=0.5.1",
    "soundfile>=0.13.1",
    "websocket-client>=1.8.0",
]
